+++ src/Components/Shared/Shared.jsx
@@ -6,12 +6,13 @@ import { Link, useNavigate } from "react-router-dom";
 const Shared = () => {
   const navigate = useNavigate(); //useNavigate allows programmatic navigation
   return (
-    <div class="navbar">
-        <ul>
-          <li class="navlinks"><Link to="/"> Home </Link></li>
-          <li class="navlinks"><Link to="/about"> About </Link></li>
-          <li class="navlinks"><Link to="/bands"> Bands </Link></li>
-          <li class="navlinks"><Link to="/venues"> Venues </Link></li>
+    <div className="navbar">
+        <ul >
+          <li className="navlinks"><Link to="/"> Home </Link></li>
+          <li className="navlinks"><Link to="/about"> About </Link></li>
+          <li className="navlinks"><Link to="/bands"> Bands </Link></li>
+          <li className="navlinks"><Link to="/venues"> Venues </Link></li>
+          <li className="navlinks"><Link to="/loggedIn"> User </Link></li>
         </ul>
     </div>
   );