+++ CHANGELOG.md
@@ -3,7 +3,30 @@ All notable changes to this project will be documented in this file.
  
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
+## [0.3.0] - 2025-04-01
+
+ ## [0.2.2] - 2025-04-01
  
+### Added
+    Added Auth Modules
+    Added Logout button to mainlist component
+### Changed
+    Updated auth for band registration
+    Modified authform for band name and genre
+    Modified authservice to save band name
+### Fixed
+    Fixed venues components to corretly display data
+    fixed protected route to not allow unauthorized users into the loggedIn page
+
+## [0.2.1] - 2025-03-30
+ 
+### Added
+    Added protected route
+### Changed
+    Updated routing 
+### Fixed
+    
+
 ## [0.2.0] - 2025-03-06
  
 ### Added